试图解决Client端释放Server端的对象问题

+++ como/runtime/type/comostring.h
@@ -84,7 +84,7 @@ public:
 
     Integer GetByteLength() const;
 
-    Integer GetHashCode() const;
+    Long GetHashCode() const;
 
     inline const char* string() const;
